Refactor: UIColorExtension 코드를 ViewController로 이동 및 리턴타입을 Double로 변경
