--- DrawingApp/DrawingApp/UIColorExtension.swift
@@ -1,20 +1,0 @@
-//
-//  UIColorExtension.swift
-//  DrawingApp
-//
-//  Created by Bibi on 2022/03/17.
-//
-
-import UIKit
-
-extension UIColor {
-    var rgba: (red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat) {
-        var red: CGFloat = 0
-        var green: CGFloat = 0
-        var blue: CGFloat = 0
-        var alpha: CGFloat = 0
-        getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-
-        return (red, green, blue, alpha)
-    }
-}
